+++ packages/create-dara-app/create_dara_app/templates/default/{{ cookiecutter.__project_slug }}/{{ cookiecutter.__package_name }}/main.py
@@ -5,7 +5,6 @@ from {{cookiecutter.__package_name}}.utils.template import template_renderer
 from dara.core import ConfigurationBuilder
 from dara.core.css import get_icon
 
-
 # Create the configuration builder
 config = ConfigurationBuilder()