+++ packages/dara-components/dara/components/common/base_component.py
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Optional, Literal
+from typing import Literal, Optional
 
 from dara.core.definitions import ComponentInstance, StyledComponentInstance